set up the environment of development (#2)
This branch has the environment docker for
the project. It will be the base
environment for the project, but you can
also use it as a sandbox when learning
react or nest. You can look at the schema
in docs/assets/docker_env_dev.png.

This branch has the environment docker for
the project.  It will be the base
environment for the project, but you can
also use it as a sandbox when learning
react or nest.  You can look at the schema
in docs/assets/docker_env_dev.png.