database ressources added entities updated
Channel with null pass OK, date in database
Date is no longer included in Channel Type(front) because is directly assigned in TypeOrm (in the back).
Id is no sended also because it's a primary Key so TypeOrm will assign the value even if i send directly un id number.