Avoiding malicious data in the requests (#35)
* Avoiding malicious data in the requests

By adding the option 'whitelist: true'
to 'ValidationPipe' we eliminate all
the unexpected data that may be sent
to us in HTTP requests.

https://docs.nestjs.com/techniques/validation#stripping-properties

* Bugfix

Bug: Could not add a new user

Explanation:

This behavior was due to the
ValidationPipe's "whitelist" parameter.

When this parameter is activated
(whitelist: true), the ValidationPipe
automatically discards any property not
decorated by validators in the DTO class.

This means that if a property has no
validation decorator, it will be
considered invalid and removed from the
object during the validation process.

Solution:

Adding @Allow() decorator to the
"avatarUrl" field. By doing this we'are
telling the ValidationPipe to allow the
field to be present and not perform any
specific validation on it. This way,
the "avatarUrl" property will not be
discarded by the ValidationPipe.