This is a combination of 12 commits
* This is the 1st commit message:

Installation of the optilib library

Otplib is a JavaScript One Time Password
(OTP) library for OTP generation and
verification that we will use in 2fa

cmd: npm install otplib

https://github.com/yeojz/otplib#about

* This is the commit message #2:

Generate 2fa secrets

auth.service.ts:

- create
generateTwoFactorAuthenticationSecret()
method. This method aims to generate the
secret and otpAuthUrl in the
authentication service and return both of
them. The 'COSMIC_PONG’  is the name that
will appear in the google authenticator
app

player.module.ts:

- export the TypOrmModule module so that
the auth module can use it

player.service.ts:

- creation of the
setTwoFactorAuthenticationSecret() method
to update the player in the database with
the secret just generated

player.ts:

- added column
'twoFactorAuthenticationSecret' to the
Player entity to store player secret

* This is the commit message #3:

qrcode package installation

Qrcode is a barcode and qrcode generator
that we will use to transform the
otpAuthUrl (generated by the otplib
library) into qrcode to be later scanned
by the authenticated google application.

Command used: npm install --save qrcode

Doc: https://github.com/soldair/node-qrcode

* This is the commit message #4:

Generate qr

auth.service.ts:

- Added generateQrCodeDataURL() method to
generate the qr code based on the
otpAuthUrl previously generated by otplib

make format

* This is the commit message #5:

Add two-factor authentication functionality

Player.ts:

- added isTwoFactorAuthenticationEnabled
column to entity

player.service.ts:

- added turnOnTwoFactorAuthentication and
turnOffTwoFactorAuthentication methods

auth.service.ts:

- addition of the
isTwoFactorAuthenticationCodeValid()
method that verifies the authentication
code using the player's secret.

* This is the commit message #6:

Add the turn on route in auth.controller

Feature not finished, still in progress

* This is the commit message #7:

Installation of the strategy TOTP

Installation of the Passport strategy for
two-factor authentication using a TOTP
value

Docs:
- https://github.com/jaredhanson/passport-totp
- http://www.passportjs.org/packages/passport-totp

Comando:
- npm install passport-totp

* This is the commit message #8:

Creation of 2fa authentication endpoint

Creation of 2fa guard and strategy

Added comments explaining passport
functionalities

* This is the commit message #9:

temporal comment to patch bug

* This is the commit message #10:

Conflict resolution during rebasing

Upgrade from 'Player' to 'User' in
all services, modules, classes, etc.

Adaptation of code to current
structure that did not exist at the
time of creating this branch.

* This is the commit message #11:

Creation of frontend pages

Frontend:

- Creation of TFAVerify and TFAActivate
pages

- TwoFactorVerificationBox is only a
mockup at the moment. It does not interact
with the backend

Backend:

- GET auth/2fa/generate endpoint
creation

- This endpoint uses two
services from the 'auth' module, which in
turn uses services from the 'user'
module

- generateTwoFactorAuthenticationSecret()
creates a TOTP secret and assigns it to
the user

- generateQrCodeDataURL()
generates a QR code that is forwarded to
the front end as a response

* This is the commit message #12:

Creation of TFA pages

Frontend:

TFAAuthenticate.tsx:
- Creation of page that will serve to
authenticate the user using the
‘TwoFactorVerificationBox' element.

TFATurnOn.tsx:
- Creation of page that will serve for
the user to enable the 2FA. It shows a QR
code to scan with google authenticator.
It also provides a
‘TwoFactorVerificationBox' element for the
user to send his code to the backend for
verification purposes.

TwoFactorVerificationBox.tsx:
- Creation of element
'TwoFactorVerificationBox' which
will be in charge of receiving the code
entered by the user and send it as
follows: POST auth/2fa/authenticate

Backend:

auth.controller.ts
(Endpoint POST auth/2fa/authenticate):
- Accept and process the request
verifying that the code that the
TwoFactorVerificationBox sent in the
body of the request is correct.

Auth42.strategy.ts:
- Temporary added functionality that
creates the ME user in the database.
This functionality is being worked on
in the s_create_player branch.