Profil page fetch user infos
* The profile page now fetches the
following user information from the
database:
- login
- avatarUrl
- nbVictory
- totalPlay
- xp
- TFAEnabled

The following values are calculated in
the frontend as follows:

- Level = nbVictory / 5 + 1
- Defeats = totalPlay - nbVictory

Next feature: Ranking

It will be calculated by sorting in
ascending order the 'users' table taking
into account the amount of experience.
Then the user's position in this table will
be counted and this value will be used
as ranking

The TypeOrm request to use will look like
this:
const userPosition = await this.userRepository
.createQueryBuilder('user')
.where('user.xp <= :userXp', { userXp:
user.xp }).getCount()

* Other changes made in this commit:
Removed private user data that was sent
as a response when a GET api/auth/status
request was made.