Enhance 2FA Handling and Session Flag Update
auth.controller.ts:
- @Post('2fa/authenticate')
If the TOTP provided is valid, the
req.session.needTFA flag is changed to
false so that the 'authenticated.guard' is
passed successfully.

authenticated.guard.ts:
- Now also controls that the user does
not need to authenticate with the 2fa.

oauth.guard.ts:
- The value of the user variable
'TFAEnabled' is copied to a
session variable called 'needTFA'.
Subsequently 'needTFA' will change its
value from 'true' to 'false' if the 2fa
authentication is successful. the 2fa
authentication succeeds.