2fa (#71)
New feature available 🎉 

From now on, the user can activate the use 
of 2fa directly from his profile page. 

When the user clicks to activate it, he is 
redirected to the page localhost:
4040/tfaturnOn . There he is provided with 
a QR code that he must scan with a TOTP 
code generation application and then send 
the verification code it generated to our 
server. If the code is correct, the user's 
TFAEnabled attribute is changed to true 
and will be taken into account the next 
time he/she tries to log in. 

The page localhost:4040/tfaturnOn should 
ideally not be accessible if the user has 
already enabled 2fa so if you have ideas 
on how to implement it, they are welcome. 

Most of our endpoints will be protected by 
the 'AuthenticatedGuard' guard. So now 
AuthGuard checks two things. Firstly, that 
the user has successfully logged in with 
42. Secondly, it verifies that the user 
has successfully executed the 2fa (if in 
its attributes in the database TFAEnabled 
=== true ). 

There is also the page localhost:
4040/tfaverify . This page should be shown 
to a user who (having previously and 
successfully activated 2fa) has 
successfully logged in with his login 42. (
If you have ideas how to make these 
redirections, they are welcome).This page 
exists for the user to send his 
verification code to the backend for 
verification and if it is correct, he is 
allowed to login to his profile.This page 
should not be accessible once the user has 
entered their code correctly. From the 
backend, the endpoint related to this page 
is already protected taking into account 
the last consideration but not yet in the 
frontend (if you have ideas on how to 
implement it, they are welcome). 

Note that the creation of the user in our 
database (once the user successfully logs 
in with 42) is still random. Each time we 
login with 42 a random user is created. 
This feature is being worked on in 
s_create_player. 

To test the new feature you can download 
your favorite TOTP creation application 
and try to enable and disable 2fa.