Creation of /user/nickname page (#74)
Both in the frontend and in the backend.

This page will be used in the Chat
section. More specifically in each
'friend' icon that according to the
subject has to be able to allow us to
visit the profile of another user.

It is a basic page but I think it meets
the requirements. If you want to add more
information it is always possible.

If you want to access a url with the
nickname of a non-existent user,
simply redirect the traffic to /profile.

Modify the Dummy Data to have varied
values in the user experiences in order to
have a realistic ranking.

To test it, you can:

    make clean-database
    make
    GET localhost:8080/api (to create
    dummy data)
    Login with 42
    Go to localhost:4040/user/user1
    Go to localhost:4040/user/user2
    Go to localhost:4040/user/user3
    Go to localhost:4040/user/user4
    Go to localhost:4040/user/user5
    Go to localhost:4040/user/user6
    Go to localhost:4040/user/your_nickname
    Go to localhost:4040/user/non-existent_nickname