Improved text box
Now the text box only appears when we
click on the button. Also, after entering
a text, pressing the enter key executes
the editProfile() method to send the text
to the backend. Also, the text is deleted
in the frontend to leave the box free for
a next request. And finally, the text box
is closed.