Change nickname
Now the user can change his nickname by
pressing the pencil icon next to his name.
Then enter a text and press enter.

If the user provides a nickname that
already exists in the database he will
receive an error.

If he provides an empty string, he will
receive an error.

If he provides an available nickname, it
will be updated in the database and the
frontend page will be reloaded showing the
new name.

This functionality is not explicitly
protected against SQL injections as no
user data is sanitized. It only validates
your input taking into account that it is
not an empty string. As far as I
understand TypeORM does a great job
protecting us against SQL injections. If
more security is needed we can implement
it.