2fa redirection
Implementation of routing based on actual
login status.

Being on the SignIn page, if the user
accesses successfully with login 42 and
does not have 2fa enabled, he is
redirected to the profile page.

Being on the SignIn page, if the user
successfully logs in with login 42 and has
2fa enabled, he/she will be redirected to
the 2fa verification page. If he/she
enters a correct code, he/she is
redirected to the profile page.

If the user is already fully authenticated
but tries to log in to the SignIn page,
he/she is redirected to the profile page.

Added logout button (need to improve the
css because it is not readable when
hovering the mouse over the button).This
button is available in the Navbar and in
the 2fa verification page.