2fa bug fix
Sometimes, the code sent by the user was
not valid.

Cause: the endpoint running the QR code
creation method was not properly
protected. This meant that if the user
entered localhost:4040/tfaturnOn several
times, new QR codes were generated.

Solution: protect the endpoint. If the
user already has 2fa enabled, a new QR
will not be generated and displayed even
if the user tries to access the address
localhost:4040/tfaturnOn.

Thanks @RobinBurri for reporting it.