2fa frontend routing (#81)
* 2fa redirection

Implementation of routing based on actual
login status.

Being on the SignIn page, if the user
accesses successfully with login 42 and
does not have 2fa enabled, he is
redirected to the profile page.

Being on the SignIn page, if the user
successfully logs in with login 42 and has
2fa enabled, he/she will be redirected to
the 2fa verification page. If he/she
enters a correct code, he/she is
redirected to the profile page.

If the user is already fully authenticated
but tries to log in to the SignIn page,
he/she is redirected to the profile page.

Added logout button (need to improve the
css because it is not readable when
hovering the mouse over the button).This
button is available in the Navbar and in
the 2fa verification page.

Co-authored-by: |°_°| <grivel.theo@protonmail.com>

* 2fa bug fix

Sometimes, the code sent by the user was
not valid.

Cause: the endpoint running the QR code
creation method was not properly
protected. This meant that if the user
entered localhost:4040/tfaturnOn several
times, new QR codes were generated.

Solution: protect the endpoint. If the
user already has 2fa enabled, a new QR
will not be generated and displayed even
if the user tries to access the address
localhost:4040/tfaturnOn.

Thanks @RobinBurri for reporting it.

* Typo

Victorires -> Victories

---------

Co-authored-by: fire-poy <99355290+fire-poy@users.noreply.github.com>
Co-authored-by: Emiliano Presa custodio <epresa-c@c3r4s5.42lausanne.ch>
Co-authored-by: |°_°| <grivel.theo@protonmail.com>