UserLambda bugfix
The <UserLambda> element was protected by
<ProtectedRoute> in the frontend and this
gave problems.

Solution: remove the use of
<ProtectedRoute> in the frontend and
instead, use the AuthenticatedGuard in the
backend to protect the endpoint where
UserLambda does the FETCH. This way, if
the user is authenticated correctly,
UserLambda will be able to FETCH
information from the back. If the user is
not authenticated, <UserLambda> will not
succeed in doing FETCH and consequently
will not be able to display any
information.

This is a hotfix, since in the future the
AuthenticatedGuard will be set by default
to all endpoints.