Global usage of authenticated.guard (#90)
The 'authenticated.guard' guard is now 
used globally throughout the backend. 

If you want it not to be used on a certain 
endpoint you must use the @Public 
decorator. 

@Public is for not using the global guard 
called authenticated.guard. 

@UseGuards(Activate2faGuard), for example, 
is to use a different guard called 
Activate2faGuard which verifies that the 
user is able to activate the 2fa. Namely, 
the user must be logged in and must not 
have already activated the 2fa. 

I have added @Public in two occasions: 

- In the first endpoints that we access 
when we enter the site for the first time 
and that do not require us to be logged 
in. 

- In cases where there are other guards 
involved. 

Docs:https:
//docs.nestjs.com/security/authentication#enable-authentication-globally