Save match in DB, improvement collisions, leave room handling (#109)
* Saving of matches in the database

In the format [ player_one , player_two ]

Next commit:
- change the format to [ winner , loser ]

* Saving of matches in the database

Format change to [ winner , loser ]

* Clean some TODOs

* improved coalition logic

* bugfix and clean unused code

the bug caused 0 0 to be saved in the match history score

* Rooms now react if a player leaves

- Room ID now passed along with User ID
when leaving a room in pong.gateway.ts.
Reduces reliance on array-like access of
rooms, mitigating 'undefined' room issues.

- Logic change: Client leaving a room now
broadcasts 'leftRoom' event to the room,
not just the leaving client.

- Comment corrected from "COALITION
LOGIQUE" to "COALITION LOGIC".

- Unnecessary comments and logs removed in
BoardGame.tsx and Play.tsx to clean up
codebase.

- Added 'otherPlayerHasLeftTheRoom' ref to
handle scenario where another player
leaves the room in BoardGame.tsx.

- Implemented a new 'leftRoom' event
listener in BoardGame.tsx. On triggering,
it notifies user and navigates back to
'/play' after a delay.

* typo

* typo

* Improved Socket Handling

- Removed unnecessary console.log
statements from the RoomService  and
BoardGame components.

- Optimized the deletion of rooms from the
matching system.

- Refactored socket.close() usage in the
BoardGame component.  This was causing
issues with event emission. the next step
is to instantiate the socket globally so
that it is always open.

- Simplified the rendering logic in the
main.tsx file.

* delete console.log